+++ utils/benchmark.py
@@ -7,6 +7,7 @@ import numpy as np
 import pandas as pd
 import pytablewriter
 from stable_baselines3.common.results_plotter import load_results, ts2xy
+
 from utils import get_trained_models
 
 parser = argparse.ArgumentParser()