+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## Pre-Release 0.9.0a2 (WIP)
+
+### Breaking Changes
+
+### New Features
+
+### Bug fixes
+- Fixed `TimeFeatureWrapper` inferring max timesteps
+
+### Documentation
+
+### Other
+- Added tests for the wrappers
+
+
 ## Release 0.8.0 (2020-08-04)
 
 ### Breaking Changes